+++ main/pages/stocks.php
@@ -92,7 +92,7 @@
     <h1 class="h2"> <span class="text-muted">Stocks</span></h1>
     <div class="btn-toolbar mb-2 mb-md-0">
       <div class="h5 mr-5">
-        <i class="fa fa-user mr-1"></i> Welcome: <?=$_SESSION["name"];?>
+      <i class="fa fa-user mr-1"></i> Welcome: <?=$_SESSION["name"];?> <?php if($_SESSION["role"]==0){echo "(Super Admin)";}else if($_SESSION["role"]==1){echo "(Property Personnel)";}else if($_SESSION["role"]==2){echo "(Laboratory Staff)";} ?>
       </div>
       <div class="h5">
         <i class="far fa-calendar mr-1"></i> <?=date("F d, Y");?>
@@ -102,11 +102,11 @@
 
   <div class="row mb-2">
     <div class="col-12">
-
+    <?php if($_SESSION["role"] == 0||$_SESSION["role"] == 1){?>
       <div class="btn-group mb-3 float-right">
         <button class="btn btn-sm btn-outline-success" data-toggle="modal" data-target="#add_stock">Add</button>
         <button class="btn btn-sm btn-outline-danger" onclick="delete_stocks()">Delete</button>
-      </div>
+      </div><?php }?>
       <div class="table-responsive">
         <table id="tbl_stocks" class="table table-striped table-bordered table-sm">
           <thead>
@@ -123,8 +123,7 @@
               <th width="100">Date Purchase</th>
               <th width="100">Repair Sched</th>
               <th width="100">Status</th>
-              <th width="80">Actions</th>
-         
+              <?php if($_SESSION["role"] == 0||$_SESSION["role"] == 1){?> <th width="100">Action</th> <?php } ?>   
             </tr>
           </thead>
           <tbody>
@@ -357,7 +356,7 @@
 <!-- PAGE SCRIPT -->
 <script type="text/javascript">
   $(document).ready( function(){
-    get_stocks();
+    get_stocks(<?php echo $_SESSION['role']?>);
     $('.custom-select').select2();
 
     $(function() {
@@ -384,8 +383,56 @@
     }
   }
 
-  function get_stocks(){
-    $("#tbl_stocks").DataTable().destroy();
+  function get_stocks(val){
+    if(val==2){
+      $("#tbl_stocks").DataTable().destroy();
+    $("#tbl_stocks").dataTable({
+        dom: 'Bfrtip',
+        buttons: [
+        'excel',
+    ],
+      "ajax": {
+        "type": "POST",
+        "url": "../ajax/datatables/stocks_data.php",
+      },
+      "processing": true,
+      "columns": [
+      {
+        "mRender": function(data, type, row){
+          return "<input type='checkbox' value='"+row.stock_id+"' name='cb_stock'>";
+        }
+      },
+      {
+        "data": "count"
+      },
+      {
+        "data": "product"
+      },
+      {
+        "data": "supplier_name"
+      },
+      {
+        "data": "engine_number"
+      },
+      {
+        "data": "location_name"
+      },
+      {
+        "data": "cost_price"
+      },
+      {
+        "data": "expiry_date"
+      },
+      {
+        "data": "date_repair"
+      }, {
+        "data": "status"
+      }
+      ]
+
+    });
+    }else{
+      $("#tbl_stocks").DataTable().destroy();
     $("#tbl_stocks").dataTable({
         dom: 'Bfrtip',
         buttons: [
@@ -436,6 +483,8 @@
       ]
 
     });
+    }
+   
   }
 
   $("#add_stock_form").submit( function(e){
@@ -459,7 +508,7 @@
               $("input").val("");
               $("textarea").html("");
               $("select").val(0).trigger('change');
-              get_stocks();
+              get_stocks(1);
             }else{
               alert("Error: "+data);
               $("#add_stock").modal("hide");
@@ -533,7 +582,7 @@
           $("#edit_stock").modal("hide");
           $("input").val("");
           $("textarea").html("");
-          get_stocks();
+          get_stocks(1);
         }else{
           alert("Error: "+data);
         }
@@ -560,7 +609,7 @@
           success: function(data){
             if(data != 0){
               alert("Success! Selected Product Stocks was deleted.");
-              get_stocks();
+              get_stocks(1);
             }else{
               alert("Error: "+data);
             }
